+++ packages/expo-modules-core/android/src/main/java/expo/modules/kotlin/methods/TypeInformation.kt
@@ -1,0 +1,3 @@
+package expo.modules.kotlin.methods
+
+data class TypeInformation<T>(val type: Class<T>, val isNullable: Boolean)