+++ packages/expo-modules-core/android/src/main/java/expo/modules/kotlin/modules/ModuleDefinition.kt
@@ -1,0 +1,9 @@
+package expo.modules.kotlin.modules
+
+import expo.modules.kotlin.methods.AnyMethod
+
+class ModuleDefinition(
+  val name: String,
+  val constantsProvider: () -> Map<String, Any?>,
+  val methods: Map<String, AnyMethod>
+)